Minor Changes Change to Date and Time Format Change in Logger Package and Class Name
