+++ src/main/resources/application.properties
@@ -7,4 +7,6 @@ spring.datasource.username=root
 spring.datasource.password=1234
 logging.file.name=app.log
 logging.level.root=INFO
+logging.pattern.console=%date{dd-MM-yyyy | HH:mm:ss} [%p] [%t] - %msg%n
+logging.pattern.file=%date{dd-MM-yyyy | HH:mm:ss} [%p] [%t] - %msg%n
 logging.level.org.springframework.aop=DEBUG
