+++ src/test/java/com/aspire/employee_api_v3/service/EmployeeApiServiceTest.java
@@ -332,15 +332,15 @@ public class EmployeeApiServiceTest {
         when(employeeJpaRepository.findById(1)).thenReturn(Optional.of(mockEmployee));
         when(accountJpaRepository.findByName("Aspire Machine Learning")).thenReturn(Optional.of(mockAccount));
         when(streamJpaRepository.findById("IND-ASP-ML-DELIVERY")).thenReturn(Optional.of(mockStream));
-        when(employeeJpaRepository.existsByManagerId(1)).thenReturn(true); 
-        
+        when(employeeJpaRepository.findByManagerId(1)).thenReturn(List.of(mockEmployee1));
+
         CustomException exception = assertThrows(CustomException.class, () -> {
             employeeApiService.updateEmployeeAccountName(1,"Aspire Machine Learning","IND-ASP-ML-DELIVERY");;
-        }); 
+        });
 
         assertThat(exception.getMessage()).isEqualTo("Account name of a manager with subbordinates can't be updated");
 
-    
+
     }
 
     @Test
@@ -352,7 +352,7 @@ public class EmployeeApiServiceTest {
         when(employeeJpaRepository.findById(1)).thenReturn(Optional.of(mockEmployee));
         when(accountJpaRepository.findByName("Aspire Machine Learning")).thenReturn(Optional.of(mockAccount));
         when(streamJpaRepository.findById("IND-ASP-ML-DELIVERY")).thenReturn(Optional.of(mockStream));
-        when(employeeJpaRepository.existsByManagerId(1)).thenReturn(false); 
+        when(employeeJpaRepository.findByManagerId(1)).thenReturn(Collections.emptyList());
         when(employeeJpaRepository.findByStreamAndManagerId(mockStream, 0)).thenReturn(streamManager);
